test(task-migrate-clean-batch): 迁 handedness_guard 7 用例到 gui_unit_test
420.5 Step 5，owner 裁决 D5：判定「重要 ⇒ 直接搬」，不暂停。

该文件是三条独立正向投影实现（backend lm_proj / GUI 预览着色器镜像 /
GUI 标签放置）之间的跨层一致性 oracle，直接 include 6 个 core/ 与 config/ 头
——doc/api-layering-and-product-lines.md 记的「唯一未守住的 API 边界」违规方。
gui_unit_test 是全仓唯一同时链 lumice_gui_obj + lumice_obj 的 target，
故为该 oracle 唯一可容纳处；搬完后 test/gui/ 下不再有 core/ 直连。

- 新增 test/unit-correctness/gui/test_render_handedness_guard.cpp（7 个 TEST）
- 删除 test/gui/functional/test_render_handedness_guard.cpp + 注册声明/调用点
- 文件头改写：把「permitted because it lives under test/gui/」换成
  「这是跨层 oracle，gui_unit_test 是唯一能做这次比较的 target」；
  顺带把指向 git-ignored 工作制品的引用改写为机制本身

核验：断言 IM_CHECK 30 → EXPECT_TRUE 30 一一对应；globe 三条与 single-lens
三条方向相反（设计使然）逐条核对未"顺手统一"；gui_test 548→541；
gui_unit_test 27→34 全绿。
